Refactor window/rendering from minifb to winit + pixels

+++ simple-engine/src/assets/tileset.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 pub struct Tile {
-    pub pixels: Vec<u32>,
+    pub pixels: Vec<u8>, // raw RGBA bytes, 4 bytes per pixel
 }
 
 pub struct Tileset {
